pyproject.toml
[project]
name = "tldbotto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp~=3.10.11",
    "app-store-server-library>=1.9.0",
    "apscheduler~=3.7.0",
    "arrow~=1.3.0",
    "asyncache>=0.3.1",
    "cachetools>=5.5.2",
    "discord-py>=2",
    "dnspython>=2.7.0",
    "emoji==1.2.0",
    "pyjwt[crypto]>=2.10.1",
    "pymongo>=4.14.1",
    "python-dateutil~=2.8.1",
    "pytz~=2021.1",
    "yarl~=1.12.1",
]

[dependency-groups]
dev = [
  "pytest~=6.2.4"
]
